Move sklearn related classes to new module clustering.sklearn_clustering
