+++ src/sensai/clustering/__init__.py
@@ -1,2 +1,3 @@
 from .greedy_clustering import GreedyAgglomerativeClustering
-from sensai.clustering.clustering_base import SkLearnEuclideanClusterer, EuclideanClusterer
+from sensai.clustering.clustering_base import EuclideanClusterer
+from .sklearn_clustering import SkLearnEuclideanClusterer
