+++ src/sensai/clustering/coordinate_clustering.py
@@ -4,7 +4,9 @@ import numpy as np
 from shapely.geometry import MultiPoint
 from typing import Callable, Union, Iterable
 
-from sensai.clustering.clustering_base import EuclideanClusterer, SkLearnClustererProtocol, SkLearnEuclideanClusterer
+from sensai.clustering.clustering_base import EuclideanClusterer
+from . import SkLearnEuclideanClusterer
+from .sklearn_clustering import SkLearnClustererProtocol
 from ..util.cache import LoadSaveInterface
 from ..util.coordinates import validateCoordinates, extractCoordinatesArray, TCoordinates, GeoDataFrameWrapper
 from ..util.tracking import timed
